pyproject.toml
[tool.poetry]
name = "bdp2-haaq"
version = "0.1.0"
description = "Big Data Programming 2 - Twitter sentiment classification project"
authors = ["Hema Pandey <hemapandey.srh@gmail.com>"]
readme = "README.md"
packages = [{include = "bdp2_haaq"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
dvc = "^2.24.0"
dvc-gdrive = "^2.19.0"
blis = "0.9.1"
catalogue = "2.0.8"
certifi = "2022.9.14"
cffi = "1.15.1"
charset-normalizer = "2.1.1"
colorama = "0.4.5"
confection = "0.0.1"
configparser = "5.3.0"
cryptography = "38.0.1"
cymem = "2.0.6"
dnspython = "2.2.1"
emoji = "2.0.0"
idna = "3.4"
Jinja2 = "3.1.2"
joblib = "1.2.0"
langcodes = "3.3.0"
MarkupSafe = "2.1.1"
murmurhash = "1.0.8"
nltk = "3.7"
numpy = "1.23.3"
oauthlib = "3.2.1"
packaging = "21.3"
pandas = "1.4.4"
pathy = "0.6.2"
preshed = "3.0.7"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pydantic = "1.9.2"
pymongo = "4.2.0"
pyOpenSSL = "22.0.0"
pyparsing = "3.0.9"
python-dateutil = "2.8.2"
pytz = "2022.2.1"
regex = "2022.9.13"
requests = "2.28.1"
requests-oauthlib = "1.3.1"
service-identity = "21.1.0"
six = "1.16.0"
smart-open = "5.2.1"
srsly = "2.4.4"
thinc = "8.1.1"
tqdm = "4.64.1"
tweepy = "4.10.1"
typer = "0.4.2"
typing-extensions = "4.3.0"
urllib3 = "1.26.12"
wasabi = "0.10.1"
absl-py = "^1.2.0"
tensorflow = "^2.10.0"
dvc-s3 = "^2.20.1"
sklearn = "^0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
